+++ src/platforms/github.spec.ts
@@ -56,6 +56,7 @@ function createTestConfig(): Config {
     streamingEnabled: true,
     streamingLines: 5,
     tempPath: "./.mergementor",
+    tokenSaver: false,
   };
 }